Incorporate a test JSON loader which is only present during tests.
Keeping the JSON loader in montague_testapps up to date was quite a hassle, and it was really the only cause of changes to that package. So I've moved it here.

Unfortunately, making a given entry point be present only during a test was a bit harder, especially since a lot of pkg_resources works only with a global working_set. Finally got it working, though.